+++ Src/SOS.Process/Processors/VirtualHerbarium/VirtualHerbariumObservationFactory.cs
@@ -88,7 +88,7 @@ namespace SOS.Process.Processors.VirtualHerbarium
                 },
                 Identification = new Identification
                 {
-                    UncertainDetermination = false,
+                    UncertainIdentification = false,
                     Validated = false,
                     ValidationStatus = new VocabularyValue { Id = (int)ValidationStatusId.ReportedByExpert }
                 },